refactor(editor): split Skins.cpp into one file per skin (audit #109 F005)
Editor/skins/Skins.cpp had grown to 2258 lines holding all five built-in skins
(studio, minimal, soft, rack, matrix), their private paint/QSS helpers, the
matrix board widgets, and the registry. Audit #109 finding F005 flagged it as too
large to navigate.

Split it verbatim into one translation unit per skin plus a shared header:

  - StudioSkin.cpp / MinimalSkin.cpp / SoftSkin.cpp / RackSkin.cpp /
    MatrixSkin.cpp each hold that skin's helpers and class, moved unchanged.
  - SkinSupport.h holds finishTokens() (the one helper every skin shared, now an
    inline) and declares the per-skin factory accessors.
  - Each skin exposes a process-lifetime instance through xSkin(); the former
    file-scope g_* globals are gone.
  - Skins.cpp keeps only Skins::all()/byId(), now assembling the roster from the
    accessors. all() order and byId() aliases (glassy->studio, industrial->rack)
    are unchanged.

This is a pure move: no helper logic, QSS, paint code, or skin behaviour changed,
so the rendered skins are identical. Verified with a clean Editor qmake/nmake
build: all five new skin translation units plus the Editor link cleanly. CI also
builds the Editor (qmake/nmake) on every PR.

No version bump (refactor: - internal file structure only, no user-visible change).
