+++ Editor/skins/RackSkin.cpp
@@ -1,0 +1,200 @@
+/*
+	This file is part of EqualizerAPO-XT, a system-wide equalizer.
+*/
+
+// Rack skin, split out of Skins.cpp (audit #109 F005). This is a verbatim
+// move of the helpers and the class; behaviour is unchanged. The file-scope
+// instance is exposed through rackSkin() so Skins::all() can assemble the
+// roster without a central definition list.
+
+#include "Skins.h"
+
+#include <QAction>
+#include <QComboBox>
+#include <QDial>
+#include <QEvent>
+#include <QFontMetrics>
+#include <QFontMetricsF>
+#include <QHBoxLayout>
+#include <QIcon>
+#include <QLabel>
+#include <QLayout>
+#include <QPainter>
+#include <QPainterPath>
+#include <QPixmap>
+#include <QStyle>
+#include <QToolBar>
+#include <QVBoxLayout>
+#include <QWidget>
+#include <QtMath>
+
+// Studio's S3 band-colour law maps BiQuad filter types onto hue families.
+#include "filters/BiQuad.h"
+
+#include "Editor/SkinManager.h"
+#include "Editor/helpers/GUIHelper.h"
+#include "Editor/skins/RackChrome.h"
+#include "Editor/skins/pickers/StudioFilterPicker.h"
+#include "Editor/skins/pickers/MinimalFilterPicker.h"
+#include "Editor/skins/pickers/SoftFilterPicker.h"
+#include "Editor/skins/pickers/RackFilterPicker.h"
+#include "Editor/skins/pickers/MatrixFilterPicker.h"
+#include "Editor/widgets/routing/CrosspointMatrixRoutingRenderer.h"
+#include "Editor/widgets/routing/StepListRoutingRenderer.h"
+#include "Editor/widgets/routing/BlockChipRoutingRenderer.h"
+#include "Editor/widgets/routing/CurvedNodeRoutingRenderer.h"
+#include "Editor/widgets/routing/HardwarePatchbayRoutingRenderer.h"
+#include "SkinSupport.h"
+
+namespace
+{
+class RackSkin : public ISkin
+{
+public:
+	QString id() const override { return QStringLiteral("rack"); }
+	QString qssResource(bool dark) const override
+	{
+		return QStringLiteral(":/skins/rack_%1.qss").arg(dark ? QStringLiteral("dark") : QStringLiteral("light"));
+	}
+	IRoutingRenderer* routingRenderer() const override
+	{
+		static HardwarePatchbayRoutingRenderer renderer;
+		return &renderer;
+	}
+
+	void paintKnob(QPainter& painter, const QRect& rect, const KnobState& state, const SkinTokens& tokens) const override
+	{
+		RackChrome::paintKnob(painter, rect, state, tokens);
+	}
+
+	QString cardFrameStyle(const CommandRowInfo& info, const SkinTokens& tokens) const override
+	{
+		// QSS only provides the machined base plate and the hover brightening;
+		// the faceplate texture, ears, screws and LEDs are painted on top by
+		// RackChrome::paintCardChrome (the sheen overlays are translucent, so
+		// the hover state shines through them). The resting border is the dark
+		// seam of the rack opening rather than the token border, so stacked
+		// units separate physically (R3); focus and selection keep their
+		// signal colours.
+		const bool dark = QColor(tokens.background).lightness() < 128;
+		const QString seam = dark ? QStringLiteral("#060809") : QStringLiteral("#8F8268");
+		const QString borderColor = info.focused ? tokens.focusRing : (info.selected ? tokens.accent : seam);
+		const QString background = info.selected ? tokens.cardSelected : tokens.card;
+		return QStringLiteral(
+			"QFrame#FilterCardRow { background: %1; border: 1px solid %2; border-radius: %3px; }"
+			"QFrame#FilterCardRow:hover { background: %4; }")
+			.arg(background, borderColor)
+			.arg(tokens.borderRadius)
+			.arg(info.selected ? tokens.cardSelected : tokens.cardHover);
+	}
+
+	QString cardHeaderStyle(const CommandRowInfo&, const SkinTokens&) const override
+	{
+		// The header strip is part of the painted faceplate; a transparent
+		// background lets the brushed metal, ears and LEDs show through.
+		return QStringLiteral("QWidget#FilterCardHeader { background: transparent; }");
+	}
+
+	void prepareCommandRow(const CommandRowInfo& info, QWidget* card, QWidget* header, QWidget* body) const override
+	{
+		// Reserve the rack-ear zones along the faceplate edges so the painted
+		// chrome (screws, LEDs, patchbay jacks, the VST nameplate) never
+		// collides with row content. Rows are rebuilt on every skin switch, so
+		// this only ever runs while the rack skin is active.
+		if (header != nullptr && header->layout() != nullptr)
+		{
+			const int right = RackChrome::earWidth() + 6
+				+ (info.type == QStringLiteral("vst") ? RackChrome::nameplateReserve() : 0);
+			header->layout()->setContentsMargins(RackChrome::earWidth() + 6, 4, right, 4);
+		}
+		// Only the modern card's body stack is inset; body-only consultations
+		// (Include/VST editors, legacy rows) already sit inside that stack.
+		if (card != nullptr && body != nullptr)
+			body->setContentsMargins(RackChrome::earWidth() + 4, 0, RackChrome::earWidth() + 4, 6);
+	}
+
+	void paintCardChrome(QPainter& painter, const QRect& rect, const CommandRowInfo& info, const SkinTokens& tokens) const override
+	{
+		RackChrome::paintCardChrome(painter, rect, info, tokens);
+	}
+
+	void paintTitleBarChrome(QPainter& painter, const QRect& rect, const SkinTokens& tokens) const override
+	{
+		// The caption strip is the unit's top panel: brushed sheen, machined
+		// edges, the caption-ear groove and two rail screws (RackChrome). QSS
+		// prints the model designation and dresses the caption buttons as
+		// machined caps.
+		RackChrome::paintTitleBarChrome(painter, rect, tokens);
+	}
+
+	FilterPickerView* createFilterPicker(QWidget* parent) const override
+	{
+		// The module library browser: a brushed 1U faceplate with engraved
+		// section plates, LED-lit slots and an LCD search strip.
+		return new RackFilterPickerView(parent);
+	}
+
+	void styleMainToolbar(QToolBar* toolBar, const SkinTokens& tokens) const override
+	{
+		// The shared stroke icons first (tinted with the panel's warm ink),
+		// then the master-rail chrome: RackChrome mounts a painted overlay
+		// (brushed strip, machined edges, end screws, engraved series
+		// marking, instant-mode power LED) under the toolbar's controls. The
+		// QSS dresses the controls themselves as transport buttons and an
+		// LCD save-state well.
+		ISkin::styleMainToolbar(toolBar, tokens);
+		RackChrome::styleMainToolbar(toolBar, tokens);
+	}
+
+	SkinTokens tokens(bool dark) const override
+	{
+		SkinTokens t;
+		t.fontFamily = QStringLiteral("DM Sans");
+		t.monoFontFamily = QStringLiteral("DM Mono");
+		// Machined plate corners; raw config lines stay off the faceplate (the
+		// "..." raw editor still reaches them) - hardware prints no raw text.
+		t.borderRadius = 3;
+		t.showRawPreview = false;
+		t.rowHeight = 36;
+		t.channelGroupIndent = 16;
+		t.channelGroupStyle = SkinTokens::DottedLine;
+		t.badgeStyle = SkinTokens::WireframeBorder;
+		t.accent = dark ? QStringLiteral("#F4B860") : QStringLiteral("#B66A00");
+		t.accent2 = dark ? QStringLiteral("#5ED0A0") : QStringLiteral("#177A55");
+		if (dark)
+		{
+			t.background = QStringLiteral("#0B0D0F");
+			t.surface = QStringLiteral("#14181C");
+			t.card = QStringLiteral("#1D2328");
+			t.cardHover = QStringLiteral("#252B2F");
+			t.cardSelected = QStringLiteral("#332718");
+			t.text = QStringLiteral("#E6E0D4");
+			t.mutedText = QStringLiteral("#9A9488");
+			t.border = QStringLiteral("#3A4248");
+			t.graph = QStringLiteral("#060807");
+			t.graphGridMinor = QStringLiteral("#1F3A31");
+		}
+		else
+		{
+			t.background = QStringLiteral("#E7E2D8");
+			t.surface = QStringLiteral("#F4EFE5");
+			t.card = QStringLiteral("#FFFAEF");
+			t.cardHover = QStringLiteral("#F7EEDC");
+			t.cardSelected = QStringLiteral("#FCE8BD");
+			t.text = QStringLiteral("#2B2721");
+			t.mutedText = QStringLiteral("#746A5D");
+			t.border = QStringLiteral("#C9BFAE");
+			t.graph = QStringLiteral("#FFF7E6");
+			t.graphGridMinor = QStringLiteral("#D6C4A6");
+		}
+		finishTokens(t);
+		return t;
+	}
+};
+}
+
+ISkin* rackSkin()
+{
+	static RackSkin instance;
+	return &instance;
+}
